Reduce size of PR
